requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
unstructured>=0.18.0
selenium>=4.21.0
chromadb>=0.5.5
streamlit>=1.35.0
pandas>=2.2.0
python-dotenv>=1.0.0